@@ -2,7 +2,7 @@ source 'https://rubygems.org'
 ruby '2.3.1'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.7'
+gem 'rails', '5.0.0'
 gem 'responders', '~> 2.1'
 
 # Caching
@@ -112,7 +112,3 @@ end
 
 # Use unicorn as the app server
 # gem 'unicorn'
-
-# Use debugger
-# gem 'debugger', group: [:development, :test]
-#
