@@ -1,4 +1,4 @@
-class Benefit < ActiveRecord::Base
+class Benefit < ApplicationRecord
   belongs_to :user
 
   validates :message, presence: true
