@@ -1,4 +1,4 @@
-class CreditCardDecline < ActiveRecord::Base
+class CreditCardDecline < ApplicationRecord
   belongs_to :user
   validates :stripe_fingerprint, presence: true
 end
