@@ -1,4 +1,4 @@
-class Message < ActiveRecord::Base
+class Message < ApplicationRecord
   belongs_to :user
   belongs_to :target_user, class_name: 'User'
   belongs_to :dialogue
