@@ -1,4 +1,4 @@
-class Refund < ActiveRecord::Base
+class Refund < ApplicationRecord
   serialize :metadata, Hash
 
   belongs_to :payment
