@@ -1,4 +1,4 @@
-class StripeTransfer < ActiveRecord::Base
+class StripeTransfer < ApplicationRecord
   serialize :stripe_response, Hash
   belongs_to :user
   validates :user_id, :amount, presence: true
